Create ChristmasModelDecisionToPurchase
Extension to the Christmas Utility Function, that takes the utility output, and determines when people will go shopping based on the rule: If the cost to go shopping today is the lowest, you go. I compare costs incurred under present bias and under EDU to get an understanding of the magnitude to which this affects people. The agent table is a good way to see for individuals in the model what the parameter values impact.