Create ErdosRenyi
A basic function to create a numpy array which can be interpreted as an adjacency matrix of a Erdos-Renyi network. For those unfamiliar with ER networks, they are (arguably) the simplest random network, whereby the probability of an edge existing between any two nodes is consistent across all nodes in the network. My function allows functionality to parameterise whether symmetry (i.e. if an edge exists between nodes i and j, then an edge will also exist from node j to i- the real-world analogy of an asymmetric network would be on twitter 'following someone', and a symmetric network would be on facebook when two people are 'friends') is possible. I'm not sure whether symmetry is allowed in a pure ER network, but whatever.